upload files fixed

+++ admin/blogs/delete_blogs.php
@@ -2,9 +2,19 @@
 
 if (!empty($_GET["d"])) {
 
-  $sql = "DELETE FROM blogs WHERE id_blog = " . $_GET["d"];
+  $sql = "SELECT * FROM blogs WHERE id_blog = " . $_GET["d"];
+  $result = $conn->query($sql);
+  if ($result->num_rows > 0) {
+    $row = $result->fetch_assoc();
+    echo $thumbnail_url = $row["thumbnail_url"];
 
-  $conn->query($sql);
+    // delete thumbnail
+    unlink("../../blog/img/$thumbnail_url");
 
-  header("location:{$url}admin/blogs/");
+    // delete from database
+    $sql = "DELETE FROM blogs WHERE id_blog = " . $_GET["d"];
+    $conn->query($sql);
+  }
 }
+
+header("location:{$url}admin/blogs/");
